refactor:refactor daily closing form and details (#114)
